Updating CONTRIBUTING.md to reflect proper gulp command arguments. (#12)
In the CONTRIBUTING.md file, under the "Package Tests" section, we had written that...

"...you can `gulp build --project odo-carousel` or even `gulp build -p carousel`."

However, the `get-package-directory.js` file seems to be referencing the argument `--package` and not `--project`. The Error that is thrown when you type the command wrong also references `--package`.

This updates the docs to reflect the correct argument for running a gulp command from the root directory.